+++ packages/autotool/examples/01_node/CHANGELOG.md
@@ -1,5 +1,21 @@
 # @llamaindex/autotool-01-node-example
 
+## 0.0.6
+
+### Patch Changes
+
+- Updated dependencies [4810364]
+- Updated dependencies [d3bc663]
+  - llamaindex@0.5.25
+  - @llamaindex/autotool@2.0.1
+
+## 0.0.5
+
+### Patch Changes
+
+- llamaindex@0.5.24
+- @llamaindex/autotool@2.0.1
+
 ## 0.0.4
 
 ### Patch Changes